move helper to selectors
